+++ src/features/usage-consolidation-accordion/usage-consolidation-content-resource.js
@@ -1,7 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { FormattedMessage } from 'react-intl';
 
+import FullTextRequestUsageTable from './full-text-request-usage-table';
 import SummaryTable from './summary-table';
+import NoCostPerUseAvailable from './no-cost-per-use-available';
 import {
   costPerUse as costPerUseShape,
   entityTypes,
@@ -10,14 +13,33 @@ import {
 
 const propTypes = {
   costPerUseData: costPerUseShape.CostPerUseReduxStateShape.isRequired,
-  year: PropTypes.string.isRequired,
+  platformType: PropTypes.string.isRequired,
+  startMonth: PropTypes.string.isRequired,
+  year: PropTypes.number.isRequired,
 };
 
-const UsageConsolidationContentResource = props => {
-  const data = props.costPerUseData.data[costPerUseTypes.RESOURCE_COST_PER_USE];
-
+const UsageConsolidationContentResource = ({
+  costPerUseData,
+  platformType,
+  startMonth,
+  year,
+  ...rest
+}) => {
+  const data = costPerUseData.data[costPerUseTypes.RESOURCE_COST_PER_USE];
   if (!data) {
     return null;
+  }
+
+  const { isFailed } = costPerUseData;
+
+  if (isFailed) {
+    return (
+      <div data-test-cost-per-use-request-is-failed>
+        <FormattedMessage
+          id="ui-eholdings.usageConsolidation.fullTextRequestUsageTable.noResponse"
+        />
+      </div>
+    );
   }
 
   const {
@@ -32,16 +54,30 @@ const UsageConsolidationContentResource = props => {
     columnMapping: { cost: 'ui-eholdings.usageConsolidation.summary.resourceCost' },
   };
 
-  return (
-    <SummaryTable
-      id="resourceUsageConsolidationSummary"
-      entityType={entityTypes.RESOURCE}
-      customProperties={customProperties}
-      noCostPerUseAvailable={noCostPerUseAvailable}
-      costPerUseType={costPerUseTypes.PACKAGE_COST_PER_USE}
-      {...props}
-    />
-  );
+  return noCostPerUseAvailable
+    ? (
+      <NoCostPerUseAvailable
+        entityType={entityTypes.RESOURCE}
+        year={year}
+      />
+    ) : (
+      <>
+        <SummaryTable
+          id="resourceUsageConsolidationSummary"
+          entityType={entityTypes.RESOURCE}
+          customProperties={customProperties}
+          noCostPerUseAvailable={noCostPerUseAvailable}
+          costPerUseType={costPerUseTypes.RESOURCE_COST_PER_USE}
+          costPerUseData={costPerUseData}
+          {...rest}
+        />
+        <FullTextRequestUsageTable
+          usageData={data.attributes.usage}
+          platformType={platformType}
+          startMonth={startMonth}
+        />
+      </>
+    );
 };
 
 UsageConsolidationContentResource.propTypes = propTypes;