+++ src/features/usage-consolidation-accordion/usage-consolidation-content-title.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import {
   useIntl,
@@ -11,6 +11,8 @@ import {
   List,
 } from '@folio/stripes/components';
 
+import FullTextRequestUsageTable from './full-text-request-usage-table';
+import NoCostPerUseAvailable from './no-cost-per-use-available';
 import SummaryTable from './summary-table';
 import {
   formatCoverageYear,
@@ -29,23 +31,37 @@ import styles from './usage-consolidation-content.css';
 
 const propTypes = {
   costPerUseData: costPerUseShape.CostPerUseReduxStateShape.isRequired,
+  platformType: PropTypes.string.isRequired,
   publicationType: PropTypes.string,
+  startMonth: PropTypes.string.isRequired,
   year: PropTypes.string.isRequired,
 };
 
-const UsageConsolidationContentTitle = (props) => {
+const UsageConsolidationContentTitle = ({
+  costPerUseData,
+  platformType,
+  startMonth,
+  publicationType,
+  year,
+}) => {
   const intl = useIntl();
-  const { costPerUseData, publicationType } = props;
-  const [sortParameters, onHeaderClick] = useMultiColumnListSort(sortOrders.asc, 'packageName');
-  const {
-    sortOrder,
-    sortedColumn,
-  } = sortParameters;
+  const [{ sortOrder, sortedColumn }, onHeaderClick] = useMultiColumnListSort(sortOrders.asc, 'packageName');
 
   const data = costPerUseData.data[costPerUseTypes.TITLE_COST_PER_USE];
-
   if (!data) {
     return null;
+  }
+
+  const { isFailed } = costPerUseData;
+
+  if (isFailed) {
+    return (
+      <div data-test-cost-per-use-request-is-failed>
+        <FormattedMessage
+          id="ui-eholdings.usageConsolidation.fullTextRequestUsageTable.noResponse"
+        />
+      </div>
+    );
   }
 
   const holdingsSummary = data?.attributes?.analysis?.holdingsSummary;
@@ -145,19 +161,31 @@ const UsageConsolidationContentTitle = (props) => {
     formatter,
   };
 
-  return (
-    <SummaryTable
-      id="titleUsageConsolidationSummary"
-      contentData={contentData}
+  return noCostPerUseAvailable ? (
+    <NoCostPerUseAvailable
       entityType={entityTypes.TITLE}
-      customProperties={customProperties}
-      noCostPerUseAvailable={noCostPerUseAvailable}
-      costPerUseType={costPerUseTypes.TITLE_COST_PER_USE}
-      onHeaderClick={onHeaderClick}
-      sortedColumn={sortedColumn}
-      sortDirection={sortOrder.name}
-      {...props}
+      year={year}
     />
+  ) : (
+    <>
+      <SummaryTable
+        id="titleUsageConsolidationSummary"
+        contentData={contentData}
+        entityType={entityTypes.TITLE}
+        customProperties={customProperties}
+        noCostPerUseAvailable={noCostPerUseAvailable}
+        costPerUseType={costPerUseTypes.TITLE_COST_PER_USE}
+        onHeaderClick={onHeaderClick}
+        sortedColumn={sortedColumn}
+        sortDirection={sortOrder.fullName}
+        costPerUseData={costPerUseData}
+      />
+      <FullTextRequestUsageTable
+        usageData={data.attributes.usage}
+        platformType={platformType}
+        startMonth={startMonth}
+      />
+    </>
   );
 };