+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 * Title Record: Display Holdings summary table. (UIEH-999)
 * Add info popover for Usage Consolidation Accordion. (UIEH-993)
 * Edit Package/Resource record: Field labels text size is too small. (UIEH-1003)
+* Add Full text request usage table to Resource/Title Show page. (UIEH-953, UIEH-997)
+* Add Usage Consolidation Titles table to Package Show page. (UIEH-948)
 
 ## [5.0.0] (https://github.com/folio-org/ui-eholdings/tree/v5.0.0) (2020-10-15)